+++ test/wrappers/TierCalculationLibWrapper.sol
@@ -15,7 +15,7 @@ contract TierCalculationLibWrapper {
         return result;
     }
 
-    function estimatedClaimCount(uint8 _numberOfTiers, uint32 _grandPrizePeriod) external pure returns (uint32) {
+    function estimatedClaimCount(uint8 _numberOfTiers, uint16 _grandPrizePeriod) external pure returns (uint32) {
         uint32 result = TierCalculationLib.estimatedClaimCount(_numberOfTiers, _grandPrizePeriod);
         return result;
     }
